fix recurring reminders i think

+++ src/util/setReminder.ts
@@ -1,10 +1,11 @@
 import { ColorResolvable, EmbedBuilder, TextChannel } from "discord.js";
-
 import ExtendedClient from "../classes/ExtendedClient";
+
+import { emojis as emoji } from "../../config.json";
 
 import { Reminder } from "../models/Reminder";
 
-export default async function (reminder: Reminder, client: ExtendedClient): Promise<Boolean> {
+export default async function setReminder(reminder: Reminder, client: ExtendedClient): Promise<Boolean> {
     const delay = Number(Number(reminder.reminder_set) + reminder.delay) - Date.now();
 
     if (delay > client.config.reminders.timeTillSet) return false;
@@ -12,17 +13,32 @@ export default async function (reminder: Reminder, client: ExtendedClient): Prom
     client.reminders.set(
         `${reminder.user}-${reminder.reminder_id}`,
         setTimeout(async () => {
-            await reminder.deleteOne();
             client.reminders.delete(`${reminder.user}-${reminder.reminder_id}`);
+
+            if (reminder?.recurring) {
+                reminder.reminder_set = Date.now().toString();
+                await setReminder(reminder, client);
+                await reminder.save();
+            } else {
+                await reminder.deleteOne();
+            }
 
             const embed = new EmbedBuilder()
                 .setColor(client.config.embeds.default as ColorResolvable)
                 .setTitle("Reminder")
                 .setDescription(reminder.reason)
-                .addFields({
-                    name: "Set",
-                    value: `<t:${reminder.reminder_set.toString().slice(0, -3)}:f> (<t:${reminder.reminder_set.toString().slice(0, -3)}:R>)`
-                })
+                .addFields(
+                    {
+                        name: "Set",
+                        value: `<t:${reminder.reminder_set.toString().slice(0, -3)}:f> (<t:${reminder.reminder_set
+                            .toString()
+                            .slice(0, -3)}:R>)`
+                    },
+                    {
+                        name: "Recurring",
+                        value: reminder?.recurring ? emoji.tick : emoji.cross
+                    }
+                )
                 .setFooter({ text: `ID: ${reminder.reminder_id}` })
                 .setTimestamp();
